[[DOCS]] README proposal (#4)
* [[DOCS]] README proposal

This is a simplified proposal. I find it easier to understand it if there is a simple example and then the doc for each different runner (that can still be improved). The name "runner" might be confusing (I don't know if step or block is clearer). Feel free to close it if you prefer the other version.

* PR feedback